Minor bugs correction in the model

+++ Reciplease/Model/Recipe.swift
@@ -29,8 +29,7 @@ struct Recipe: Codable {
     let image: String?
     let url: String?
     let ingredients: [Ingredient]?
-    //let calories, totalWeight: Double?
-    let totalTime: Int?
+    let totalTime: Double?
 }
 
 // MARK: - Ingredient
